Time: 15 ms (74.15%), Space: 10.2 MB (94.67%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -3,7 +3,7 @@ public:
     int m=(int)1e9+7;
     int change(int amount, vector<int>& coins) {
         
-        vector<unsigned long long> dp(amount+1,0);
+        vector<unsigned int> dp(amount+1,0);
 
         for(int i=0;i<=amount;i++){
             if(i%coins[0] == 0)dp[i]=1;
@@ -12,8 +12,8 @@ public:
         for(int ind=1;ind<coins.size();ind++){
             for(int tar=0;tar<=amount;tar++){
                 
-                unsigned long long np=dp[tar];
-                unsigned long long p=0;
+                unsigned int np=dp[tar];
+                unsigned int p=0;
                 if(coins[ind] <= tar){
                     p=dp[tar-coins[ind]];
                 }
